refactor(layout): Externalise layout settings into tagged services
Refactored the monolithic KinglyLayoutBase class to delegate form
building, submission, and build processing for layout settings to a
series of dedicated, tagged services.

- Created a KinglyLayoutsDisplayOptionInterface to define the contract for
  these services.
- Created individual services for each settings category (Spacing, Color,
  Background, etc.), each implementing the interface.
- Tagged each service with \`kingly_layouts.display_option\`.
- KinglyLayoutBase now injects all tagged services and iterates over them
  to build the form, handle submission, and process the render array.
- Created a central OptionsService to provide lists of choices (e.g.,
  color palettes, spacing scales) to the other services, reducing
  code duplication.

This change improves maintainability, separates concerns, and makes it
easier to add or modify layout setting categories in the future."
