+++ kingly_layouts.services.yml
@@ -1,7 +1,57 @@
 services:
+  kingly_layouts.options:
+    class: Drupal\kingly_layouts\Service\OptionsService
+    arguments: ['@entity_type.manager', '@cache.default', '@string_translation']
   kingly_layouts.spacing:
     class: Drupal\kingly_layouts\Service\SpacingService
-    arguments: [ '@current_user', '@string_translation' ]
+    arguments: ['@current_user', '@string_translation']
+    tags:
+      - { name: kingly_layouts.display_option }
   kingly_layouts.alignment:
     class: Drupal\kingly_layouts\Service\AlignmentService
-    arguments: [ '@current_user', '@string_translation' ]
+    arguments: ['@current_user', '@string_translation', '@kingly_layouts.options']
+    tags:
+      - { name: kingly_layouts.display_option }
+  kingly_layouts.color:
+    class: Drupal\kingly_layouts\Service\ColorService
+    arguments: ['@current_user', '@string_translation', '@kingly_layouts.options', '@entity_type.manager']
+    tags:
+      - { name: kingly_layouts.display_option }
+  kingly_layouts.border:
+    class: Drupal\kingly_layouts\Service\BorderService
+    arguments: ['@current_user', '@string_translation', '@kingly_layouts.options', '@kingly_layouts.color']
+    tags:
+      - { name: kingly_layouts.display_option }
+  kingly_layouts.animation:
+    class: Drupal\kingly_layouts\Service\AnimationService
+    arguments: ['@current_user', '@string_translation', '@kingly_layouts.options']
+    tags:
+      - { name: kingly_layouts.display_option }
+  kingly_layouts.background:
+    class: Drupal\kingly_layouts\Service\BackgroundService
+    arguments: ['@current_user', '@string_translation', '@kingly_layouts.options', '@kingly_layouts.color']
+    tags:
+      - { name: kingly_layouts.display_option }
+  kingly_layouts.shadows_effects:
+    class: Drupal\kingly_layouts\Service\ShadowsEffectsService
+    arguments: ['@current_user', '@string_translation', '@kingly_layouts.options']
+    tags:
+      - { name: kingly_layouts.display_option }
+  kingly_layouts.responsiveness:
+    class: Drupal\kingly_layouts\Service\ResponsivenessService
+    arguments: ['@current_user', '@string_translation', '@kingly_layouts.options']
+    tags:
+      - { name: kingly_layouts.display_option }
+  kingly_layouts.custom_attributes:
+    class: Drupal\kingly_layouts\Service\CustomAttributesService
+    arguments: ['@current_user', '@string_translation']
+    tags:
+      - { name: kingly_layouts.display_option }
+  kingly_layouts.container_type:
+    class: Drupal\kingly_layouts\Service\ContainerTypeService
+    arguments: ['@current_user', '@string_translation', '@kingly_layouts.options']
+    tags:
+      - { name: kingly_layouts.display_option }
+  kingly_layouts.display_option_manager:
+    class: Drupal\kingly_layouts\Service\DisplayOptionManager
+    arguments: ['@container.namespaces', '@cache.default', '@module_handler']
