+++ src/Service/AlignmentService.php
@@ -23,16 +23,24 @@ class AlignmentService implements KinglyLayoutsDisplayOptionInterface {
   protected AccountInterface $currentUser;
 
   /**
+   * The options service.
+   */
+  protected OptionsService $optionsService;
+
+  /**
    * Constructs a new AlignmentService object.
    *
    * @param \Drupal\Core\Session\AccountInterface $current_user
    *   The current user.
    * @param \Drupal\Core\StringTranslation\TranslationInterface $string_translation
    *   The string translation service.
+   * @param \Drupal\kingly_layouts\Service\OptionsService $options_service
+   *   The options service.
    */
-  public function __construct(AccountInterface $current_user, TranslationInterface $string_translation) {
+  public function __construct(AccountInterface $current_user, TranslationInterface $string_translation, OptionsService $options_service) {
     $this->currentUser = $current_user;
     $this->stringTranslation = $string_translation;
+    $this->optionsService = $options_service;
   }
 
   /**
@@ -49,7 +57,7 @@ class AlignmentService implements KinglyLayoutsDisplayOptionInterface {
     $form['alignment']['vertical_alignment'] = [
       '#type' => 'select',
       '#title' => $this->t('Vertical Alignment'),
-      '#options' => $this->getVerticalAlignmentOptions(),
+      '#options' => $this->optionsService->getOptions('vertical_alignment'),
       '#default_value' => $configuration['vertical_alignment'],
       '#description' => $this->t('Align content vertically within the layout. This assumes the layout uses Flexbox or Grid. "Stretch" makes columns in the same row equal height.'),
     ];
@@ -57,7 +65,7 @@ class AlignmentService implements KinglyLayoutsDisplayOptionInterface {
     $form['alignment']['horizontal_alignment'] = [
       '#type' => 'select',
       '#title' => $this->t('Horizontal Alignment'),
-      '#options' => $this->getHorizontalAlignmentOptions(),
+      '#options' => $this->optionsService->getOptions('horizontal_alignment'),
       '#default_value' => $configuration['horizontal_alignment'],
       '#description' => $this->t('Justify content horizontally within the layout. This assumes the layout uses Flexbox or Grid.'),
     ];
@@ -86,43 +94,10 @@ class AlignmentService implements KinglyLayoutsDisplayOptionInterface {
   /**
    * {@inheritdoc}
    */
-  public function defaultConfiguration(): array {
+  public static function defaultConfiguration(): array {
     return [
       'vertical_alignment' => 'center',
       'horizontal_alignment' => 'start',
-    ];
-  }
-
-  /**
-   * Gets the options for vertical alignment.
-   *
-   * @return array
-   *   An associative array of options.
-   */
-  private function getVerticalAlignmentOptions(): array {
-    return [
-      'stretch' => $this->t('Stretch'),
-      'flex-start' => $this->t('Top'),
-      'center' => $this->t('Center (Default)'),
-      'flex-end' => $this->t('Bottom'),
-      'baseline' => $this->t('Baseline'),
-    ];
-  }
-
-  /**
-   * Gets the options for horizontal alignment.
-   *
-   * @return array
-   *   An associative array of options.
-   */
-  private function getHorizontalAlignmentOptions(): array {
-    return [
-      'start' => $this->t('Start (Left)'),
-      'center' => $this->t('Center'),
-      'end' => $this->t('End (Right)'),
-      'space-between' => $this->t('Space Between'),
-      'space-around' => $this->t('Space Around'),
-      'space-evenly' => $this->t('Space Evenly'),
     ];
   }
